Multiple revisions:
- uses Emacs built-ins for parsing ISO 8601 strings

- explicitly defines hourly and daily forecast types

Since leaning on shell command `date` begs for portability issues, it is preferable to use Emacs built-ins for parsing ISO 8601 strings.

Explicitly defining types of forecasts provides guidance regarding how a forecast set should be handled. Due to the transition to relying on the forecast-set struct, NOAA-LAST-FORECAST-SET, NOAA-CLEAR-FORECAST-SET, NOAA-POPULATE-FORECASTS are used in place of NOAA-LAST-FORECAST, NOAA-CLEAR-LAST-FORECAST, NOAA-SET-LAST-FORECAST. In addition, this commit implements NOAA-DISPLAY-LAST-FORECAST-AS-DAILY and NOAA-INSERT-HOUR-FORECAST along with separate callbacks for HTTP requests depending on whether forecast requested is daily or hourly.